% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{any_pkg_behind_capsule}
\alias{any_pkg_behind_capsule}
\title{check if any packages are behind capsule}
\usage{
any_pkg_behind_capsule(
  dep_source_paths = "./packages.R",
  lockfile_path = "./renv.lock"
)
}
\arguments{
\item{dep_source_paths}{a character vector of file paths to extract
package dependencies from.}

\item{lockfile_path}{a length one character vector path of the lockfile for}
}
\value{
TRUE if dev packages are behind capsule, FALSE otherwise.
}
\description{
A wrapper for \link{get_pkg_behind_capsule} that returns TRUE if any
dependencies found in \code{dep_source_paths} are behind the lockfile version in
\code{lockfile_path}
}
\examples{
# ADD_EXAMPLES_HERE
}
