% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capshot.R
\name{capshot}
\alias{capshot}
\alias{capshot_str}
\title{Quickly generate an renv compliant lock file}
\usage{
capshot(
  dep_source_paths = "./packages.R",
  lockfile_path = "./renv.lock",
  minify = FALSE
)

capshot_str(dep_source_paths = "./packages.R", minify = FALSE)
}
\arguments{
\item{dep_source_paths}{files to scan for project dependencies to write to the lock file.}

\item{lockfile_path}{output path for the lock file.}

\item{minify}{a boolean value indicicating if lockfile JSON should have whitespace removed to shrink footprint.}
}
\value{
\code{lockfile_path}. Writes lockfile as a side effect.
}
\description{
These functions generate json lockfiles that can be restored from using
\code{capsule} or \code{renv}.
}
\details{
Unlike \code{\link[=create]{create()}} this function does not populate a local library.
It writes a lock file using dependencies found in files in \code{dep_source_paths}.
Package dependency information is mined from DESCRIPTION files using the
current \code{\link[=.libPaths]{.libPaths()}}.

These functions do not use \code{{renv}} machinery and so may produce
different results. They have been re-implmented for speed, so that they can
be integrated into automated pipelines that build projects or documents.
}
\section{Functions}{
\itemize{
\item \code{capshot_str()}: a variation that returns lockfile json as a character vector for further use.

}}
