% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{create}
\usage{
create(dep_source_paths = "./packages.R")
}
\arguments{
\item{dep_source_paths}{files to find package dependencies in.}
}
\value{
nothing. Creates a capsule as a side effect.
}
\description{
Create a capsule library context to run code in
}
\details{
Dependencies to be encapsulated are detected from files you nominate in
\code{dep_source_paths}. Good practice would be to have a single dependencies R
file that contains all library() calls - hence this makes an explicit
assertion of your dependencies. This way spurious usages of pkg:: for
packages not stated as dependencies will cause errors that can be caught.
}
\author{
Miles McBain
}
