% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tracking.R
\name{create_renv_lockfile}
\alias{create_renv_lockfile}
\title{Create renv Lockfile}
\usage{
create_renv_lockfile(output_file = "renv.lock", project_path = ".")
}
\arguments{
\item{output_file}{Character. Path to save lockfile. Default "renv.lock".}

\item{project_path}{Character. Path to project. Default is current directory.}
}
\value{
Path to created lockfile
}
\description{
Generate an renv-compatible lockfile for package reproducibility
}
\examples{
\dontrun{
create_renv_lockfile("renv.lock")
}
}
