% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{delete}
\usage{
delete()
}
\value{
nothing
}
\description{
Delete the capsule
}
\details{
Removes the lockfile and library, in the case that you made a mistake or no
longer want to use capsule.
}
\author{
Miles McBain
}
