% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_local_lib}
\alias{delete_local_lib}
\title{delete_local_lib}
\usage{
delete_local_lib()
}
\value{
nothing
}
\description{
Delete the capsule's local library
}
\details{
This helper is provided to help you recover from mistakes or test building
the library from a lockfile you have generated.
}
\seealso{
[delete()]
}
\author{
Miles McBain
}
