% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_lockfile}
\alias{delete_lockfile}
\title{delete_lockfile}
\usage{
delete_lockfile()
}
\value{
nothing
}
\description{
Delete the capule's lockfile
}
\details{
This helper is provided to help you recover from mistakes or test extracting
dependencies.
}
\author{
Miles McBain
}
