% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_dependencies.R
\name{detect_dependencies}
\alias{detect_dependencies}
\title{Detect dependencies in nominated R or Rmd files.}
\usage{
detect_dependencies(file_path)
}
\arguments{
\item{file_path}{the file(s) to detect dependencies in.}
}
\value{
a character vector of package names
}
\description{
Get the names of R packages referred to in \code{file_path}. \code{file_path} can be a
vector of paths if you need, although I advise keeping dependency calls in a
single file for R projects.
}
\details{
This is a thin wrapper around \verb{[renv::dependencies()]} that includes support
for the \verb{[using::pkg()]} style specification via \verb{[using::detect_dependencies()]}
}
