% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{dev_mirror_lockfile}
\alias{dev_mirror_lockfile}
\title{Mirror lockfile in local library}
\usage{
dev_mirror_lockfile(
  lockfile_path = "./renv.lock",
  dep_source_paths = NULL,
  library_path = .libPaths(),
  prompt = interactive()
)
}
\arguments{
\item{lockfile_path}{lockfile to be compared to local environment.}

\item{dep_source_paths}{R files to search for dependencies, the set of packages to be updated is limited to these dependencies (and their dependencies).}

\item{library_path}{the source libary to compare versions from, and install into.}

\item{prompt}{ask for confirmation after displaying list to be installed and before isntalling?}
}
\value{
names of the packages updated or to be updated (if install did not proceed) invisibly.
}
\description{
Install packages contained in the lockfile that are either missing from the local library or at a lower version number.
}
\details{
Packages are installed at the lockfile version. Packages in the local library that are ahead of the the local library are not touched.

So this function ensures that the local development environment is \strong{at least} at the lockfile version of all packages, not \strong{equal to}.

To find differences between the local library and the lockfile versions use \code{\link[=compare_local_to_lockfile]{compare_local_to_lockfile()}}.
}
