% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_comparison.R
\name{.compare_packages}
\alias{.compare_packages}
\title{Compare Package Lists}
\usage{
.compare_packages(pkg1, pkg2)
}
\description{
Internal function to compare package versions between two snapshots
}
\keyword{internal}
