% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_comparison.R
\name{.compare_registries}
\alias{.compare_registries}
\title{Compare Registry Files}
\usage{
.compare_registries(file1, file2, type)
}
\description{
Internal function to compare registry files
}
\keyword{internal}
