% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_session.R
\name{.extract_package_info}
\alias{.extract_package_info}
\title{Extract Package Information}
\usage{
.extract_package_info(pkg_list)
}
\arguments{
\item{pkg_list}{List of package info from sessionInfo()}
}
\value{
List of package details
}
\description{
Internal function to extract detailed package information
}
\keyword{internal}
