% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_generation.R
\name{.format_r_value}
\alias{.format_r_value}
\title{Format R Value for Script}
\usage{
.format_r_value(value)
}
\arguments{
\item{value}{Any R value}
}
\value{
Character representation
}
\description{
Internal function to format R values as code strings
}
\keyword{internal}
