% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_generation.R
\name{.generate_docker_readme}
\alias{.generate_docker_readme}
\title{Generate Docker README}
\usage{
.generate_docker_readme(output_file, project_name, include_rstudio)
}
\description{
Internal function to generate Docker usage instructions
}
\keyword{internal}
