% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_generation.R
\name{.generate_dockerfile}
\alias{.generate_dockerfile}
\title{Generate Dockerfile}
\usage{
.generate_dockerfile(output_file, r_version, base_image, system_deps)
}
\description{
Internal function to generate Dockerfile content
}
\keyword{internal}
