% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_generation.R
\name{.generate_dockerignore}
\alias{.generate_dockerignore}
\title{Generate .dockerignore}
\usage{
.generate_dockerignore(output_file)
}
\description{
Internal function to generate .dockerignore file
}
\keyword{internal}
