% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tracking.R
\name{.get_fasta_stats}
\alias{.get_fasta_stats}
\title{Get FASTA Statistics}
\usage{
.get_fasta_stats(fasta_path)
}
\arguments{
\item{fasta_path}{Character. Path to FASTA file}
}
\value{
List containing FASTA statistics
}
\description{
Internal function to extract basic statistics from a FASTA file
}
\keyword{internal}
