% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_tools.R
\name{.get_tool_path}
\alias{.get_tool_path}
\title{Get Tool Path}
\usage{
.get_tool_path(tool)
}
\arguments{
\item{tool}{Character. Tool name}
}
\value{
Character. Full path or NA
}
\description{
Internal function to get the full path to a command-line tool
}
\keyword{internal}
