% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_support.R
\name{.load_conda_registry}
\alias{.load_conda_registry}
\title{Load Conda Registry}
\usage{
.load_conda_registry(registry_file)
}
\arguments{
\item{registry_file}{Character. Path to registry file}
}
\value{
List containing registry data
}
\description{
Internal function to load the conda registry
}
\keyword{internal}
