% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tracking.R
\name{.load_reference_registry}
\alias{.load_reference_registry}
\title{Load Reference Registry}
\usage{
.load_reference_registry(registry_file)
}
\arguments{
\item{registry_file}{Character. Path to registry file}
}
\value{
List containing registry data
}
\description{
Internal function to load the reference registry
}
\keyword{internal}
