% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provenance.R
\name{.load_registry}
\alias{.load_registry}
\title{Load Provenance Registry}
\usage{
.load_registry(registry_file)
}
\arguments{
\item{registry_file}{Character. Path to registry file}
}
\value{
List containing registry data
}
\description{
Internal function to load the provenance registry
}
\keyword{internal}
