% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_tools.R
\name{.load_tools_registry}
\alias{.load_tools_registry}
\title{Load Tools Registry}
\usage{
.load_tools_registry(registry_file)
}
\arguments{
\item{registry_file}{Character. Path to registry file}
}
\value{
List containing registry data
}
\description{
Internal function to load the tools registry
}
\keyword{internal}
