% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tracking.R
\name{.parse_deps}
\alias{.parse_deps}
\title{Parse Package Dependencies}
\usage{
.parse_deps(dep_string)
}
\arguments{
\item{dep_string}{Character. Dependency string from package description}
}
\value{
Character vector of package names
}
\description{
Internal function to parse package dependency strings
}
\keyword{internal}
