% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_support.R
\name{.save_conda_registry}
\alias{.save_conda_registry}
\title{Save Conda Registry}
\usage{
.save_conda_registry(registry, registry_file)
}
\arguments{
\item{registry}{List. Registry data to save}

\item{registry_file}{Character. Path to registry file}
}
\description{
Internal function to save the conda registry
}
\keyword{internal}
