% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tracking.R
\name{.save_reference_registry}
\alias{.save_reference_registry}
\title{Save Reference Registry}
\usage{
.save_reference_registry(registry, registry_file)
}
\arguments{
\item{registry}{List. Registry data to save}

\item{registry_file}{Character. Path to registry file}
}
\description{
Internal function to save the reference registry
}
\keyword{internal}
