% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provenance.R
\name{.save_registry}
\alias{.save_registry}
\title{Save Provenance Registry}
\usage{
.save_registry(registry, registry_file)
}
\arguments{
\item{registry}{List. Registry data to save}

\item{registry_file}{Character. Path to registry file}
}
\description{
Internal function to save the provenance registry
}
\keyword{internal}
