% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_tools.R
\name{.save_tools_registry}
\alias{.save_tools_registry}
\title{Save Tools Registry}
\usage{
.save_tools_registry(registry, registry_file)
}
\arguments{
\item{registry}{List. Registry data to save}

\item{registry_file}{Character. Path to registry file}
}
\description{
Internal function to save the tools registry
}
\keyword{internal}
