% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_hooks.R
\name{export_for_cwl}
\alias{export_for_cwl}
\title{Generate CWL (Common Workflow Language) Input}
\usage{
export_for_cwl(output_file = "capsule_cwl_inputs.yml")
}
\arguments{
\item{output_file}{Character. Path to save inputs. Default "capsule_cwl_inputs.yml"}
}
\value{
List containing input data
}
\description{
Export Capsule data in YAML format suitable for CWL workflows
}
\examples{
\dontrun{
export_for_cwl("reproflow_cwl_inputs.yml")
}
}
