% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_hooks.R
\name{export_for_wdl}
\alias{export_for_wdl}
\title{Create WDL (Workflow Description Language) Config}
\usage{
export_for_wdl(output_file = "capsule_inputs.json")
}
\arguments{
\item{output_file}{Character. Path to save config. Default "capsule_inputs.json"}
}
\value{
List containing config data
}
\description{
Export Capsule data in JSON format suitable for WDL workflows
}
\examples{
\dontrun{
export_for_wdl("reproflow_inputs.json")
}
}
