% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_support.R
\name{get_conda_env_info}
\alias{get_conda_env_info}
\title{Get Conda Environment Info}
\usage{
get_conda_env_info(
  env_name = NULL,
  registry_file = ".capsule/conda_registry.json"
)
}
\arguments{
\item{env_name}{Character. Specific environment name, or NULL for all}

\item{registry_file}{Character. Path to conda registry}
}
\value{
List of environment information
}
\description{
Retrieve information about tracked conda environments
}
