% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{get_pkg_behind_capsule}
\alias{get_pkg_behind_capsule}
\title{get packckes behind capsule}
\usage{
get_pkg_behind_capsule(
  dep_source_paths = "./packages.R",
  lockfile_path = "./renv.lock"
)
}
\arguments{
\item{dep_source_paths}{a character vector of file paths to extract
package dependencies from.}

\item{lockfile_path}{a length one character vector path of the lockfile for}
}
\value{
RETURN_DESCRIPTION
}
\description{
Check versions of packages referred to in your project's \code{dep_source_paths}
and identify and in your main R library that are behind the capsule
renv.lock at \code{lockfile_path}.
}
\details{
Information is returned about packages that are behind in your development
environment, so you can update them to the capsule versions if you wish.

A warning is thrown in the case that pacakges have the same version but
different remote SHA. E.g. A package in one library is from GitHub and in
the other library is from CRAN. Or Both packages are from GitHub, have the
same version but different SHAs.
}
\examples{
\dontrun{
get_pkg_behind_capsule(
  dep_source_paths = "./packages.R",
  lockfile_path = "./renv.lock"
)
}
}
