% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tracking.R
\name{list_reference_sources}
\alias{list_reference_sources}
\title{List Common Reference Genome Sources}
\usage{
list_reference_sources()
}
\description{
Display a helpful list of common reference genome sources
}
\examples{
list_reference_sources()
}
