% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate.R
\name{recreate}
\alias{recreate}
\title{recreate}
\usage{
recreate(dep_source_paths = capsule_sources())
}
\arguments{
\item{dep_source_paths}{a character vector of project source files to
extract dependencies from.  Defaults to './packages.R`.
See \code{\link[=capsule_sources]{capsule_sources()}} for setting other defaults.}
}
\value{
nothing. The capsule is regenerated as a side effect.
}
\description{
Recreate a capsule with new dependencies
}
\details{
After some development work has been completed, Use this function to update the
capsule environment to match the dependency versions in your development
environment.

Similarly to \code{create()}, you are expected to supply a vector of files in
your project to extract dependencies from. Things work best when this is a
single file containing only dependency related code.
}
\seealso{
\code{\link[=create]{create()}}
}
\author{
Miles McBain
}
