% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{repl}
\alias{repl}
\title{repl}
\usage{
repl()
}
\value{
nothing.
}
\description{
Open a REPL within the capsule
}
\details{
Uses an experimental feature from \code{callr} to attach a new process repl to
your current interactive session. That REPL evaluates code within the
context of your capsule.

To exit the process send the use the interrupt signal in the REPL e.g.
Control-C, or, ess-interrupt, or the 'stop' button in rstudio.

Depending on your R editor, overtaking your REPL with a new process may
cause strang behaviour, like the loss of autocompletions.
}
\author{
Miles McBain
}
