% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_lib}
\alias{reproduce_lib}
\title{reproduce_lib}
\usage{
reproduce_lib()
}
\value{
nothing.
}
\description{
Reproduce the capsule library from the lockfile
}
\details{
If you have cloned a project that contains a lockfile you can actually just
use \code{run()} to execute commands in the capsule and have the library built
automatically. If that is not convenient, this will explicitly create the
capsule library from the lockfile.
}
\author{
Miles McBain
}
