% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_session}
\alias{run_session}
\title{run_session}
\usage{
run_session(code)
}
\arguments{
\item{code}{an expression to run in the context of the capsule library.}
}
\value{
output of \code{code}
}
\description{
run code in the context of the capsule in the current R session
}
\details{
Execute the supplied function in the context of the capsule library, by
changing the R library paths it searches.

In almost all cases, run or run_callr which do effectively the same thing,
are preferred. This is because the \code{code} argument can cause packages
to be attached, and thus not read from the capsule library.

For example if \code{code} was \code{drake::r_make()} this would cause \code{drake}, to
compatibility issues.

Use this function when you have R code that modifies the .GlobalEnv, and
you want to inspect it at the end, or you want to actively debug with #'
browser() or recover(). Even then it may be preferrable to use
capsule::repl() to do debugging.
}
\section{Lockfile}{

At a minimum, an \code{renv} lockfile must be present in the current working
directory. The capsule library will be generated from the lockfile if it
does not exist. Use \code{create()} to make the lockfile.
}

\examples{
run(library())
run(search())
}
\seealso{
\code{\link[=create]{create()}} to make the lockfile. \code{\link[=run_callr]{run_callr()}} and \code{\link[=run]{run()}} for safer versions.
}
\author{
Miles McBain
}
