% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{whinge}
\alias{whinge}
\title{Complain if the local R library has packages that are behind the lockfile versions}
\usage{
whinge(whinge_fun = warning, lockfile_path = "./renv.lock")
}
\arguments{
\item{whinge_fun}{the function to use to have a whinge about packages, e.g. message, warning, stop, etc.}

\item{lockfile_path}{the path to the project lockfile}
}
\value{
output of whinge_fun, most likely nothing.
}
\description{
Useful for keeping teams loosely in sync on package versions. A warning can
be tolerated until updating at a convenient time. For example if
placed in the packages.R file of a \code{{tflow}} project.
}
\details{
The message is hardcoded, but the whinge_fun that takes the message is customisable.
}
